Restructuring, adding axios util file
