+++ packages/frontend/src/utils/api.ts
@@ -1,9 +1,9 @@
 import axios, { AxiosError, AxiosInstance } from 'axios';
 import createAuthRefreshInterceptor from 'axios-auth-refresh';
-import { toast } from 'react-hot-toast';
 import { getRecoil } from 'recoil-nexus';
 import { ActiveTokenSet } from '@/model/auth';
 import { requestApiAuthRefresh } from './auth';
+import { handleErrors } from './axios';
 
 /**
  * Attempt to refresh auth token and retry request
@@ -17,32 +17,6 @@ const refreshAuthTokenSet = async (err: AxiosError): Promise<void> => {
   err.response.config.headers[
     'Authorization'
   ] = `Bearer ${tokenSet.accessToken}`;
-};
-
-/**
- * Handle error responses (excluding initial 401 response)
- *
- * @param err
- */
-export const handleErrors = (err: AxiosError): void => {
-  // Any HTTP Code which is not 2xx will be considered as error
-  const statusCode = err?.response?.status;
-
-  if (err?.request && !err?.response) {
-    toast.error('Server did not respond');
-  } else if (statusCode === 404) {
-    window.location.href = '/404';
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    //@ts-ignore
-  } else if ([403, 400].includes(statusCode) && err?.response?.data?.message) {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    //@ts-ignore
-    toast.error(err.response.data.message);
-  } else if (statusCode === 401) {
-    window.location.href = '/';
-  } else {
-    toast.error('Unknown Error');
-  }
 };
 
 /**
